==> 07_final_data_set.Rmd <==
---
title: "Lifespan data set"
output:
  html_notebook:
        code_folding: hide
  pdf_document: default
---

```{r setup, include=FALSE}
current_directory <- dirname(rstudioapi::getActiveDocumentContext()$path)
knitr::opts_knit$set(root.dir=normalizePath(paste0(current_directory)))
```

```{r load_libraries, include=FALSE}
## LOAD LIBRARIES ----

library(viridis) ## colour blind friendly
library(ggpubr) ## scatter plot and correlation
library(ggridges) 
library(tidyverse) ## helpful for data handling and visualization
```

```{r}
## IMPORT DATA ----

## import CpG oe data
cpg_oe_data_unmod <- read.csv("dataFiles/06.00_cpg_oe_data.csv", row.names = 1)

## make matrix and set nas to 0
cpg_oe_data <- cpg_oe_data_unmod %>%
  ## replace NAs with zeros
  replace(is.na(.), 0) %>%
  ## format as matrix
  as.matrix(.)

## import lifespan data
lifespan_data <- read.csv("figuresTables/Table S[lifespan].csv")

## import summary metadata 
sum_metadata <- read.csv("dataFiles/06.00_metadata_sum.csv") 
```

```{r}
## create df to plot lifespan data sources for supp
lifespan_data_plot <- lifespan_data %>%
  mutate(source_group = ifelse(grepl(".,", Source), 
                               "Published articles", 
                               Source)) 

# plot
ggplot(lifespan_data_plot %>%
  group_by(source_group)  %>%
  add_count(name = "Lifespan values") %>%
  mutate(Species = n_distinct(Species)) %>%
  select(source_group, `Lifespan values`, Species) %>%
  pivot_longer(cols=c("Lifespan values", "Species"), 
               names_to = "Count type", 
               values_to = "Count") %>%
  mutate(source_group = gsub(" ", "\n", source_group)), 
       aes(source_group, 
           Count, 
           fill = `Count type`)) +   
  geom_bar(position = "dodge", 
           stat="identity")+
  labs(x=NULL)+
  geom_text(aes(source_group, label=paste0("n=", Count)),
            check_overlap = TRUE,
            size=2,
            vjust=-0.5,
            position = position_dodge(width = 1))+
  theme_bw()+
  scale_fill_viridis_d(option="C")+
  theme(legend.position = "bottom")

ggsave("figuresTables/Figure S[dataset].png", 
       width = 8.25, height = 11.75/2, units = "in", 
       device = "png")

```

```{r}
print("#### LIFESPAN DATA DISTRIBUTION ####")

ggplot(sum_metadata, 
       aes(x=mean_lifespan)) + 
  geom_histogram(colour="black",
                 fill="grey") + 
  theme_bw()

ggsave("figuresTables/pres_hist.png", 
       width = 4, height = 3, units = "in", 
       device = "png")
```


```{r}
ggplot(sum_metadata, 
       aes(x=log(mean_lifespan))) + 
  geom_histogram(colour="black",
                 fill="grey") + 
  theme_bw()

ggsave("figuresTables/pres_hist_log.png", 
       width = 4, height = 3, units = "in", 
       device = "png")
```

```{r fig.height=8}

print("#### TAXONOMIC DISTRIBUTION ####")

ggplot(sum_metadata %>% 
         select(order, genus, organism_name) %>%
         group_by(genus) %>%
         add_count(genus, name="count"), 
       aes(fill=genus, y=count, x=order)) + 
  geom_bar(position="stack", stat="identity")+
  coord_flip()+
  theme_bw()+
  theme(legend.position="none")+
  # scale_fill_grey()
  scale_fill_viridis_d(option = "C")

ggsave("figuresTables/pres_order.png", 
       width = 10, height = 6, units = "in", 
       device = "png")
```


```{r}
## PRINT SOME STATS ----

print("#### LIFESPAN ####")

## how many lifespan estimates?
paste("there are a total of", 
      length(lifespan_data$Lifespan), 
             "known lifespan values in our data set", 
             sep = " ")

## how many fish? 
paste("there are", 
      length(unique(sum_metadata$organism_name)), 
             "fish species with genomic and lifespan information in our data set", 
             sep = " ")

## how many classes are represented? 
paste("representing", 
      length(unique(sum_metadata$class)), 
             "different classes", 
             sep = " ")

## how many fish in each class
paste("number of fish in each class: ")
for (i in levels(factor(sum_metadata$class))) {
  print(paste0(i, ": ", 
               (nrow(sum_metadata %>% filter(., class == i)))))
}

## how many orders are represented? 
paste("representing", 
      length(unique(sum_metadata$order)), 
             "different orders", 
             sep = " ")

## how many families are represented? 
paste("representing", 
      length(unique(sum_metadata$family)), 
             "different families", 
             sep = " ")

## shortest lifespan
paste("the fish with the shortest lifespan is", 
      sum_metadata %>% 
        slice_min(mean_lifespan) %>%
        pull(organism_name),
      "with a mean max lifespan of", 
      round(sum_metadata %>% 
        slice_min(mean_lifespan) %>%
        pull(mean_lifespan), digits = 2), 
             "+/- SD", 
      round(sum_metadata %>% 
        slice_min(mean_lifespan) %>%
        pull(sd_lifespan), digits = 2), 
             sep = " ")

## longest lifespan
paste("the fish with the longest lifespan is", 
      sum_metadata %>% 
        slice_max(mean_lifespan) %>%
        pull(organism_name),
      "with a mean max lifespan of", 
      round(sum_metadata %>% 
        slice_max(mean_lifespan) %>%
        pull(mean_lifespan), digits = 2), 
             "+/- SD", 
      round(sum_metadata %>% 
        slice_max(mean_lifespan) %>%
        pull(sd_lifespan), digits = 2), 
             sep = " ")

## variable lifespan
paste("the fish with the most variation in reported lifespans was", 
      sum_metadata %>% 
        slice_max(sd_lifespan) %>%
        pull(organism_name),
      "with a mean max lifespan of", 
      round(sum_metadata %>% 
        slice_max(sd_lifespan) %>%
        pull(mean_lifespan), digits = 2), 
             "+/- SD", 
      round(sum_metadata %>% 
        slice_max(sd_lifespan) %>%
        pull(sd_lifespan), digits = 2),
        "ranging from: ", 
      filter(lifespan_data, 
             Species == (sum_metadata %>%
                                 slice_max(sd_lifespan) %>%
                                 pull(organism_name))) %>%
        slice_min(Lifespan) %>%
        pull(Lifespan),
      "to", 
      filter(lifespan_data,
             Species == (sum_metadata %>%
                                 slice_max(sd_lifespan) %>%
                                 pull(organism_name))) %>%
        slice_max(Lifespan) %>%
        pull(Lifespan) %>%
        unique(.),
      "years",
      sep=" ")
```

```{r}
## LIFESPAN PLOTS ----

## format data ----

## shorter and longer-lived info (from summary meta)
sum_metadata_ordered <- sum_metadata %>%
  arrange(., desc(mean_lifespan))

num_rows <- nrow(sum_metadata_ordered)

longer_lived <- sum_metadata_ordered[1:(num_rows/2), ]$organism_name
shorter_lived <- sum_metadata_ordered[(num_rows/2):num_rows, ]$organism_name

## order metadata by ls
metadata_plots <- lifespan_data %>%
  ## subset by metadata species
  filter(Species %in% sum_metadata$organism_name) %>%
  ## calculate mean ls values
  group_by(Species) %>%
  mutate(mean_lifespan = mean(Lifespan),
         sd_lifespan = sd(Lifespan)) %>%
  ## order by lifespan
  arrange(., by_group = mean_lifespan) %>%
  ## rename original organism name (Species) column
  mutate(organism_name = Species) %>%
  ## add short/med/long-lived info
  mutate(ls_cat = ifelse(organism_name %in% longer_lived, 
                         "Longer-lived species",
                         ifelse(organism_name %in% shorter_lived, 
                                "Shorter-lived species", 
                                "Unknown"))) %>%
  ## remove unknowns
  filter(., ls_cat != "Unknown") %>%
  ## order factor
  mutate(ls_cat = factor(ls_cat, levels=c("Longer-lived species", "Shorter-lived species")))

## lock in spp order by lifespan
metadata_plots$organism_name <- factor(metadata_plots$organism_name, 
                                levels=unique(metadata_plots$organism_name))

```

```{r fig.height=16.5, fig.width=11.75}
## ridge plots ----
ggplot(metadata_plots,
       aes(Lifespan, organism_name,
           label=round(Lifespan)), 
       fill="grey40")+
  geom_density_ridges_gradient(rel_min_height = 0.000001,
                               scale = 0.9,
                               alpha = 0.7)+
  geom_point(size=1, shape="|", stroke=0.5)+
  geom_point(aes(mean_lifespan, organism_name), size=1.2, shape="x", stroke=0.1)+
  ylab("Species")+
  xlab("Lifespan (years)")+
  theme_bw()+
  theme(legend.position = "none", 
        axis.text.y=element_text(size=5))+
  facet_wrap(~ls_cat, scales="free", ncol=2)

ggsave("figuresTables/Figure S[lifespan_ridge].png", 
       width = 11.75, height = 16.5*0.9, units = "in", 
       device = "png")

```

```{r}
## PRINT SOME BLAST HIT STATS ----

print("#### BLAST HITS ####")

paste0("max number of hits was ", 
       sum_metadata %>% 
         slice_max(num_hits) %>%
         pull(num_hits), 
       " hits in ", 
       sum_metadata %>%
         slice_max(num_hits) %>%
         pull(organism_name))

paste0("max number of hits (non-zebrafish) was ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_max(num_hits) %>%
         pull(num_hits), 
       " hits in ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_max(num_hits) %>%
         pull(organism_name))

paste0("min number of hits (non-zebrafish) was ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_min(num_hits) %>%
         pull(num_hits), 
       " hits in ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_min(num_hits) %>%
         pull(organism_name))

paste0("max mean hit length was ", 
       round(sum_metadata %>%
         slice_max(mean_hit_len) %>%
         pull(mean_hit_len)), 
       " bp in ", 
       sum_metadata %>%
         slice_max(mean_hit_len) %>%
         pull(organism_name))

paste0("max mean hit length (non-zebrafish) was ", 
       round(sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_max(mean_hit_len) %>%
         pull(mean_hit_len), digits=2), 
       " bp in ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_max(mean_hit_len) %>%
         pull(organism_name))

paste0("min mean hit length (non-zebrafish) was ", 
       round(sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_min(mean_hit_len) %>%
         pull(mean_hit_len), digits=2), 
       " bp in ", 
       sum_metadata %>% filter(organism_name != "Danio rerio") %>%
         slice_min(mean_hit_len) %>%
         pull(organism_name))

print("#### CpG O/E ####")

paste0("min CpG O/E value: ", round(min(cpg_oe_data), digits=4))
paste0("max CpG O/E value: ", round(max(cpg_oe_data), digits=4))
paste0("min non-zero CpG O/E value: ", round(min(cpg_oe_data[cpg_oe_data > 0]), digits=7))

```

```{r}
## PLOT NUM HITS
num_hits <- ggscatter(data = sum_metadata,
          x = "dist_from_zeb_lca", 
          y = "num_hits", 
          add = "reg.line", 
          xlab = "Distance from zebrafish LCA", 
          ylab = "Number of BLAST hits",
          shape = 21, 
          ggtheme = theme_bw(),
          fill = "grey", 
          legend = "bottom", 
          legend.title = "Known lifespan") +
 stat_cor(aes(label = paste(..r.label.., ..rr.label.., 
                            ..p.label.., sep = "~`,`~")))

# num_hits
```

```{r}
## PLOT HIT LENGTH
hit_len <- ggscatter(data = sum_metadata,
          x = "dist_from_zeb_lca", 
          y = "mean_hit_len", 
          add = "reg.line", 
          xlab = "Distance from zebrafish LCA", 
          ylab = "Average BLAST hit length (bp)",
          shape = 21, 
          ggtheme = theme_bw(), 
          fill = "grey", 
          legend = "bottom", 
          legend.title = "Known lifespan")+
  stat_cor(aes(label = paste(..r.label.., ..rr.label.., 
                             ..p.label.., sep = "~`,`~")))

# hit_len
```

```{r}
## PLOT LIFESPAN
ls <- ggscatter(data = sum_metadata,
          x = "dist_from_zeb_lca", 
          y = "mean_lifespan", 
          add = "reg.line", 
          xlab = "Distance from zebrafish LCA", 
          ylab = "Known lifespan (years)",
          shape = 21, 
          ggtheme = theme_bw(), 
          fill = "grey", 
          legend = "bottom", 
          legend.title = "Known lifespan")+
  stat_cor(aes(label = paste(..r.label.., ..rr.label.., 
                             ..p.label.., sep = "~`,`~")))

# ls
```


```{r fig.height=4.125, fig.width=8.25}
## PLOT CpG Dens

cpg_oe_df <- as.data.frame(cpg_oe_data_unmod) %>%
  rownames_to_column(., var = "organism_name") %>%
  pivot_longer(., -organism_name, values_to = "cpg_oe", names_to = "promoter_id") %>% 
  group_by(organism_name) %>%
  summarise(mean_cpg_oe = mean(cpg_oe, na.rm = TRUE)) %>%
  left_join(., sum_metadata %>% select(organism_name, organism_name, dist_from_zeb_lca, mean_lifespan, dist_from_zeb_lca), 
            by = "organism_name")

cpg_oe <- ggscatter(data = cpg_oe_df,
          x = "dist_from_zeb_lca", 
          y = "mean_cpg_oe", 
          add = "reg.line", 
          xlab = "Distance from zebrafish LCA", 
          ylab = "Average CpG O/E",
          shape = 21, 
          ggtheme = theme_bw(), 
          fill = "grey", 
          legend = "bottom", 
          legend.title = "Known lifespan")+
  stat_cor(aes(label = paste(..r.label.., ..rr.label.., 
                             ..p.label.., sep = "~`,`~")))

# cpg_oe

cpg_oe_df_more <- cpg_oe_data_unmod %>%
  rownames_to_column(., var = "organism_name") %>%
  pivot_longer(., -organism_name, 
               values_to = "cpg_oe", 
               names_to = "promoter_id") %>% 
  left_join(., sum_metadata %>% select(organism_name, organism_name, 
                                       dist_from_zeb_lca, mean_lifespan), 
            by = "organism_name")

cpg_oe_more <- ggscatter(data = cpg_oe_df_more,
          x = "dist_from_zeb_lca", 
          y = "cpg_oe", 
          add = "reg.line", 
          xlab = "Distance from zebrafish LCA", 
          ylab = "CpG O/E",
          shape = 21, 
          ggtheme = theme_bw(), 
          fill = "grey", 
          legend = "bottom", 
          legend.title = "Known lifespan")+
  stat_cor(aes(label = paste(..r.label.., ..rr.label.., 
                             ..p.label.., sep = "~`,`~")))

# cpg_oe_more

ggarrange(cpg_oe, cpg_oe_more, 
          common.legend = TRUE, 
          legend = "bottom", 
          labels = c("A", "B"))
```


```{r fig.height=8.25, fig.width=8.25}
## COMBINE PLOTS

ggarrange(num_hits, hit_len, cpg_oe, ls, 
          common.legend = TRUE, 
          legend = "bottom", 
          labels = c("A", "B", "C", "D"))

ggsave("figuresTables/Figure S[prelim_correlations].png", 
       width = 8.25, height = 8.25, units = "in", 
       device = "png")

```
```{r}
## view tree data from which the cophenetic distances where derived

## load ape
library(ape)
library(ggtree)

## read in timetree data
chordate_tree <- read.tree("dataFiles/chordata_species.nwk") 

## edit tip labels
chordate_tree$tip.label <- gsub("_", " ", chordate_tree$tip.label)

## intersecting
ls_fish_tips <- intersect(chordate_tree$tip.label, 
                          unique(sum_metadata$organism_name))

## subset tree for genome and lifespan info species
ls_fish_tree <- keep.tip(chordate_tree, ls_fish_tips)

## full tree is big so remove it
rm(chordate_tree)

## re root tree to zebras (because we used zebrafish promoters)
zebra_tree <- root(ls_fish_tree, 
                   ## specify zebra fish as the new outgroup
                   outgroup = "Danio rerio", 
                   ## resolve the new root as a bifurcating node (i.e. a fork)
                   resolve.root = TRUE) 

## make ggtree compatible with gheatmap
gg_zebra <- ggtree(zebra_tree)+
  ## add tip labels and resize them
  geom_tiplab(size = 0.6)+
  ## increase margin size
  ylim(-20, 400)+
  ## add arrow
  annotate("segment", 
           x = 1000, xend = 100, 
           y = -1.5, yend = 0,
           size=0.25,
           arrow=arrow(length = unit(0.1,"cm")),
           colour = "Black")+
  ## add text
  annotate("text", 
           x = 1050,
           y = -1,
           size= 2,
           fontface = 'italic',
           label="Danio rerio")+
  ## add rectangle
  annotate("rect", 
           xmin = 200, xmax = 600, 
           ymin = 80, ymax = 400,
           alpha = 0.2, 
           fill="grey")

gg_zebra

ggsave("figuresTables/Figure S[non-ultra_tree].png",
       width = 8.25, height = 10.25, units = "in",
       device = "png")
```
 
```{r}
counts<-sum_metadata %>%
  add_count(dist_from_zeb_lca, name="num_dist_same") %>%
  pull(num_dist_same) %>%
  unique(.)

paste0("max number of species with the same distance: ", max(counts))

paste0("total number of species: ", nrow(sum_metadata))

paste0("percentage: ", round(100/nrow(sum_metadata)*max(counts), digits=2), " %")
```
